implement filter.driver.process interface
As git-annex filterdriver command, although this name is provisional.

git annex init doesn't set it up yet, and the old smudge command is
unchanged.

It's not clear that filterdriver is better at all: git has
the same memory leak using this interface as it does using the smudge
interface. And, the filterdriver clean interface necessarily has to read
the whole input from git, can't short-circuit the write the way that the
clean filter can.