WIP pre-cache location logs for git-annex get only
This should be faster, but it actually benchmarks 4x slower.
I don't know why.

Also, limits are matched before the location log pre-cache,
so if a limit like --in is used, the location log gets read twice
due to this change. (Normally, the branch cache avoids that problem, but
the use of a feeder thread prevents that from happening in this branch.)
So that would need more work even if the performance were rescued.