alternative implementation of pipeNullSplit not using lazy bytestring
Written to try to fix a memory leak involving PINNED memory. While it
does not do that, it is in many ways a nicer approach and may be worth
integrating at some point once the memory leak is understood.

The S.copy is probably unncessary, I just wanted to make sure I didn't
have one chunk pinning other chunks of the same ByteString.

This commit was sponsored by Jochen Bartl on Patreon.